+++ src/main/java/io/kontur/eventapi/dao/mapper/DataLakeMapper.java
@@ -25,4 +25,6 @@ public interface DataLakeMapper {
     DataLake getDataLakeByObservationId(UUID observationId);
 
     Optional<DataLake> getDataLakeByExternalIdAndProvider(String externalId, String provider);
+
+    Boolean isNewPdcExposure(String externalId, String geometry);
 }