+++ src/main/java/io/kontur/eventapi/pdc/config/FeignClientConfiguration.java
@@ -3,9 +3,7 @@ package io.kontur.eventapi.pdc.config;
 import feign.auth.BasicAuthRequestInterceptor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
 public class FeignClientConfiguration {
 
     @Value("${pdc.user}")