+++ src/main/java/io/kontur/eventapi/pdc/job/PdcMapSrvSearchJob.java
@@ -1,0 +1,51 @@
+package io.kontur.eventapi.pdc.job;
+
+import io.kontur.eventapi.dao.DataLakeDao;
+import io.kontur.eventapi.entity.DataLake;
+import io.kontur.eventapi.job.AbstractJob;
+import io.kontur.eventapi.pdc.client.PdcMapSrvClient;
+import io.kontur.eventapi.pdc.converter.PdcDataLakeConverter;
+import io.micrometer.core.instrument.MeterRegistry;
+import org.springframework.stereotype.Component;
+import org.wololo.geojson.Feature;
+import org.wololo.geojson.FeatureCollection;
+import org.wololo.geojson.GeoJSONFactory;
+
+import java.util.*;
+
+import static io.kontur.eventapi.util.JsonUtil.writeJson;
+
+@Component
+public class PdcMapSrvSearchJob extends AbstractJob {
+
+    private final PdcMapSrvClient pdcMapSrvClient;
+    private final PdcDataLakeConverter pdcDataLakeConverter;
+    private final DataLakeDao dataLakeDao;
+
+    public PdcMapSrvSearchJob(MeterRegistry meterRegistry, PdcMapSrvClient pdcMapSrvClient,
+                              PdcDataLakeConverter pdcDataLakeConverter, DataLakeDao dataLakeDao) {
+        super(meterRegistry);
+        this.pdcMapSrvClient = pdcMapSrvClient;
+        this.pdcDataLakeConverter = pdcDataLakeConverter;
+        this.dataLakeDao = dataLakeDao;
+    }
+
+    @Override
+    public String getName() {
+        return "pdcMapSrv";
+    }
+
+    @Override
+    public void execute() throws Exception {
+        String geoJson = pdcMapSrvClient.getExposures();
+        FeatureCollection featureCollection = (FeatureCollection) GeoJSONFactory.create(geoJson);
+        List<DataLake> dataLakes = new ArrayList<>();
+        for (Feature feature : featureCollection.getFeatures()) {
+            String externalId = String.valueOf(feature.getProperties().get("hazard_uuid"));
+            if (dataLakeDao.isNewPdcExposure(externalId, writeJson(feature.getGeometry()))) {
+                dataLakes.add(pdcDataLakeConverter.convertExposure(feature.toString(), externalId));
+            }
+        }
+        dataLakeDao.storeDataLakes(dataLakes);
+    }
+}