+++ src/main/resources/db/changelog/v0/04-update-pdc-feed.sql
@@ -1,5 +1,5 @@
 --liquibase formatted sql
 
---changeset event-api-migrations:04-update-pdc-feed runOnChange:false
+--changeset event-api-migrations:04-update-pdc-feed runOnChange:true
 
-update feeds set providers = '{"hpSrvMag", "hpSrvSearch", "pdcSqs"}' where alias = 'pdc';
+update feeds set providers = '{"hpSrvMag", "hpSrvSearch", "pdcSqs", "pdcMapSrv"}' where alias = 'pdc-v0';